move getFieldIndex to utils
