+++ src/lib/containers/home_page/resources/Resources.tsx
@@ -2,11 +2,11 @@ import React from 'react'
 import { QueryBundleRequest } from '../../../utils/synapseTypes'
 import { SynapseConstants } from '../../../utils'
 import { Error } from '../../Error'
-import { getFieldIndex } from '../goals/Goals'
 import useGetQueryResultBundle from '../../../utils/hooks/useGetQueryResultBundle'
 import ResourcesDesktop from './Resources.Desktop'
 import ResourcesMobile from './Resources.Mobile'
 import useShowDesktop from '../../../utils/hooks/useShowDesktop'
+import { getFieldIndex } from '../../../utils/functions/queryUtils'
 
 export type ResourcesProps = {
   entityId: string
